refactor(ui): modernize component boundaries
