+++ apps/landing/src/components/Button.tsx
@@ -2,141 +2,51 @@
 
 import { Slot } from '@radix-ui/react-slot'
 import { RiLoader2Fill } from '@remixicon/react'
-import React from 'react'
-import { tv, type VariantProps } from 'tailwind-variants'
-
-import { cx, focusRing } from '@repo/shared/utils'
-
-const buttonVariants = tv({
-  base: [
-    // base
-    'relative inline-flex items-center justify-center whitespace-nowrap rounded-lg border px-3 py-2 text-center font-medium text-sm shadow-sm transition-all duration-100 ease-in-out',
-    // disabled
-    'disabled:pointer-events-none disabled:shadow-none',
-    // focus
-    focusRing
-  ],
-  variants: {
-    variant: {
-      primary: [
-        // border
-        'border-transparent',
-        // text color
-        'text-white dark:text-white',
-        // background color
-        'bg-indigo-600 dark:bg-indigo-500',
-        // hover color
-        'hover:bg-indigo-700 dark:hover:bg-indigo-400',
-        // disabled
-        'disabled:bg-indigo-100 disabled:text-indigo-400',
-        'disabled:dark:bg-indigo-800 disabled:dark:text-indigo-600'
-      ],
-      secondary: [
-        // border
-        'border-gray-300 dark:border-gray-800',
-        // text color
-        'text-gray-900 dark:text-gray-50',
-        // background color
-        'bg-white dark:bg-gray-950',
-        //hover color
-        'hover:bg-gray-50 dark:hover:bg-gray-900/60',
-        // disabled
-        'disabled:text-gray-400',
-        'disabled:dark:text-gray-600'
-      ],
-      light: [
-        // base
-        'shadow-none',
-        // border
-        'border-transparent',
-        // text color
-        'text-gray-900 dark:text-gray-50',
-        // background color
-        'bg-gray-200 dark:bg-gray-900',
-        // hover color
-        'hover:bg-gray-300/70 dark:hover:bg-gray-800/80',
-        // disabled
-        'disabled:bg-gray-100 disabled:text-gray-400',
-        'disabled:dark:bg-gray-800 disabled:dark:text-gray-600'
-      ],
-      ghost: [
-        // base
-        'shadow-none',
-        // border
-        'border-transparent',
-        // text color
-        'text-gray-900 dark:text-gray-50',
-        // hover color
-        'bg-transparent hover:bg-gray-100 dark:hover:bg-gray-800/80',
-        // disabled
-        'disabled:text-gray-400',
-        'disabled:dark:text-gray-600'
-      ],
-      destructive: [
-        // text color
-        'text-white',
-        // border
-        'border-transparent',
-        // background color
-        'bg-red-600 dark:bg-red-700',
-        // hover color
-        'hover:bg-red-700 dark:hover:bg-red-600',
-        // disabled
-        'disabled:bg-red-300 disabled:text-white',
-        'disabled:dark:bg-red-950 disabled:dark:text-red-400'
-      ]
-    }
-  },
-  defaultVariants: {
-    variant: 'primary'
-  }
-})
+import { cx } from '@repo/shared/utils'
+import type { ComponentPropsWithoutRef, Ref } from 'react'
+import type { VariantProps } from 'tailwind-variants'
+import { buttonVariants } from './buttonVariants'
 
 interface ButtonProps
-  extends React.ComponentPropsWithoutRef<'button'>,
+  extends ComponentPropsWithoutRef<'button'>,
     VariantProps<typeof buttonVariants> {
   asChild?: boolean
   isLoading?: boolean
   loadingText?: string
+  ref?: Ref<HTMLButtonElement>
 }
 
-const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
-  (
-    {
-      asChild,
-      isLoading = false,
-      loadingText,
-      className,
-      disabled,
-      variant,
-      children,
-      ...props
-    }: ButtonProps,
-    forwardedRef
-  ) => {
-    const Component = asChild ? Slot : 'button'
-    return (
-      <Component
-        className={cx(buttonVariants({ variant }), className)}
-        disabled={disabled || isLoading}
-        ref={forwardedRef}
-        tremor-id="tremor-raw"
-        {...props}
-      >
-        {isLoading ? (
-          <span className="pointer-events-none flex shrink-0 items-center justify-center gap-1.5">
-            <RiLoader2Fill aria-hidden="true" className="size-4 shrink-0 animate-spin" />
-            <span className="sr-only">{loadingText ? loadingText : 'Loading'}</span>
-            {loadingText ? loadingText : children}
-          </span>
-        ) : (
-          children
-        )}
-      </Component>
-    )
-  }
-)
+function Button({
+  asChild,
+  isLoading = false,
+  loadingText,
+  className,
+  disabled,
+  variant,
+  children,
+  ref,
+  ...props
+}: ButtonProps) {
+  const Component = asChild ? Slot : 'button'
+  return (
+    <Component
+      className={cx(buttonVariants({ variant }), className)}
+      disabled={disabled || isLoading}
+      ref={ref}
+      tremor-id="tremor-raw"
+      {...props}
+    >
+      {isLoading ? (
+        <span className="pointer-events-none flex shrink-0 items-center justify-center gap-1.5">
+          <RiLoader2Fill aria-hidden="true" className="size-4 shrink-0 animate-spin" />
+          <span className="sr-only">{loadingText ? loadingText : 'Loading'}</span>
+          {loadingText ? loadingText : children}
+        </span>
+      ) : (
+        children
+      )}
+    </Component>
+  )
+}
 
-Button.displayName = 'Button'
-
-export { Button, buttonVariants, type ButtonProps }
+export { Button, type ButtonProps }
