+++ apps/landing/src/components/pricing/countdownDiff.ts
@@ -1,0 +1,10 @@
+export function diff(targetMs: number, nowMs: number) {
+  const ms = Math.max(0, targetMs - nowMs)
+  return {
+    days: Math.floor(ms / 86_400_000),
+    hours: Math.floor((ms % 86_400_000) / 3_600_000),
+    minutes: Math.floor((ms % 3_600_000) / 60_000),
+    seconds: Math.floor((ms % 60_000) / 1000),
+    expired: ms === 0
+  }
+}
